requirements.txt
pyyaml>=6.0
setuptools>=65.5.0
colorama>=0.4.6